+++ website3.0/package.json
@@ -14,27 +14,27 @@
     "@fortawesome/free-regular-svg-icons": "^6.5.2",
     "@fortawesome/free-solid-svg-icons": "^6.5.2",
     "@fortawesome/react-fontawesome": "^0.2.2",
+    "@react-oauth/google": "^0.12.1",
     "@splidejs/react-splide": "^0.7.12",
     "@splidejs/splide-extension-auto-scroll": "^0.5.3",
     "@tsparticles/all": "^3.4.0",
     "@tsparticles/react": "^3.0.0",
     "aos": "^3.0.0-beta.6",
     "axios": "^1.7.2",
+    "bcrypt": "^5.1.1",
     "clsx": "^2.1.1",
     "email-verifier": "^0.4.1",
     "framer-motion": "^11.2.11",
     "mongodb": "^6.7.0",
     "mongoose": "^8.4.4",
     "next": "14.2.4",
+    "next-auth": "^4.24.7",
     "nodemailer": "^6.9.14",
     "react": "^18",
     "react-dom": "^18",
+    "react-icons": "^5.2.1",
     "rellax": "^1.12.1",
-     "next-auth": "^4.24.7",
-    "nodemailer": "^6.9.14",
-        "bcrypt": "^5.1.1",
     "resend": "^3.4.0",
-"react-icons": "^5.2.1",
     "showdown": "^2.1.0",
     "tailwind-merge": "^2.3.0"
   },
