+++ website3.0/src/app/layout.js
@@ -4,6 +4,7 @@ import Loader from "@components/Loader";
 import Header from "@components/Header";
 import Footer from "@components/Footer";
 import Authprovider from "@components/auth/Authprovider";
+import { NextAuthProvider } from "./provider";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
@@ -18,13 +19,13 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={inter.className}>
-        <Authprovider>
+<NextAuthProvider>
 
         <Loader />
         <Header />
         {children}
         <Footer />
-        </Authprovider>
+</NextAuthProvider>
       </body>
     </html>
   );
