+++ website3.0/src/app/provider.js
@@ -1,0 +1,7 @@
+"use client"
+
+import { SessionProvider } from "next-auth/react";
+
+export const NextAuthProvider = ({ children }) => {
+  return <SessionProvider>{children}</SessionProvider>;
+};
